Performed a refactor of all basic types to type definitions in stdint.h

+++ NM3_Driver/io.cpp
@@ -1,7 +1,7 @@
 #include "io.h"
 
-void PrintChars(char* start, unsigned length) {
-	for (unsigned i = 0; i < length; i++) {
+void PrintChars(char* start, uint64_t length) {
+	for (uint64_t i = 0; i < length; i++) {
 		std::cout << *(start + i);
 	}
 }
